rest_api_test_framework/requirements.txt
allure-pytest==2.6.0
allure-python-commons==2.6.0
configparser==3.7.4
dictdiffer==0.8.0
jsonpath-rw==1.4.0
pytest==4.4.1
pytest-html==1.20.0
pytest-logger==0.4.0
pytest-metadata==1.8.0
pytest-testconfig==0.1.1
pytest-xdist==1.28.0
requests==2.18.4
urllib3==1.22
